requirements.txt
PyYAML~=6.0.1
roboflow~=1.1.24
numpy~=1.23.5
opencv-python~=4.8.0.74
supervision~=0.20.0
ipython~=7.34.0
ultralytics~=8.1.29
pandas~=2.0.3
python-dotenv~=1.0.0
scikit-learn~=1.3.0
streamlit~=1.32.2